inserting group chat session & hangout url
When the Join Hangouts button is clicked the course id & user id are
passed to the hangout as a JSON object. That JSON object is than decoded
and & the hangout URL is then added to it. An AJAX call is then made
sending all of these parameters back to lms299 web2py to insert the group
chat session into the database.